lib/kaba update

+++ src/lib/kaba/lib/lib_base.cpp
@@ -121,7 +121,7 @@ string kaba_float_format(float f, const string &fmt) {
 	}
 }
 
-string kaba_float642str(float f) {
+string kaba_float642str(double f) {
 	return f642s(f, 6);
 }
